Rename project Prototypes to prototype_FunkyButton
