+++ prototypes/prototype_FunkyButton/mainwindow.ui
@@ -76,20 +76,20 @@
      </rect>
     </property>
     <property name="text">
-     <string>gfs</string>
+     <string>Text</string>
     </property>
    </widget>
    <widget class="QPushButton" name="pushButton">
     <property name="geometry">
      <rect>
-      <x>160</x>
+      <x>150</x>
       <y>110</y>
       <width>85</width>
       <height>27</height>
      </rect>
     </property>
     <property name="text">
-     <string>PushButton</string>
+     <string>Set Colour</string>
     </property>
    </widget>
   </widget>
